+++ tests/testthat/test-snapshot.R
@@ -169,22 +169,18 @@ test_that("hint is informative", {
   local_mocked_bindings(in_check_reporter = function() FALSE)
   withr::local_envvar(GITHUB_ACTIONS = "false", TESTTHAT_WD = NA)
 
-  expect_snapshot({
-    snapshot_hint(NULL, "_default", "bar.R", reset_output = FALSE)
-    snapshot_hint(NULL, "foo", "bar.R", reset_output = FALSE)
-  })
+  expect_snapshot(snapshot_hint(NULL, "bar.R", reset_output = FALSE))
 })
 
 test_that("hint includes path when WD is different", {
   local_mocked_bindings(in_check_reporter = function() FALSE)
   withr::local_envvar(TESTTHAT_WD = "..")
 
-  hint <- snapshot_hint(NULL, "_default", "bar.R", reset_output = FALSE)
-  hint <- gsub(getwd(), "<WD>", hint)
+  hint <- snapshot_hint(NULL, "bar.R", reset_output = FALSE)
   # Can't use snapshot here because its hint will get the wrong path
   expect_match(
     hint,
-    'snapshot_accept("bar.R", "<WD>")',
+    'snapshot_accept("bar.R", "testthat")',
     fixed = TRUE,
     all = FALSE
   )